// src/app/modules/suppliers/enums/supplier.enum.ts
export enum SupplierActionTypes {
  FETCH_SUPPLIER = '[SUPPLIERS] FETCH_SUPPLIER',
  FETCH_SUPPLIER_SUCCESS = '[SUPPLIERS] FETCH_SUPPLIER_SUCCESS',
  FETCH_SUPPLIER_FAILURE = '[SUPPLIERS] FETCH_SUPPLIER_FAILURE',
  CREATE_SUPPLIER = '[SUPPLIERS] CREATE_SUPPLIER',
  CREATE_SUPPLIER_SUCCESS = '[SUPPLIERS] CREATE_SUPPLIER_SUCCESS',
  CREATE_SUPPLIER_FAILURE = '[SUPPLIERS] CREATE_SUPPLIER_FAILURE',
  UPDATE_SUPPLIER = '[SUPPLIERS] UPDATE_SUPPLIER',
  UPDATE_SUPPLIER_SUCCESS = '[SUPPLIERS] UPDATE_SUPPLIER_SUCCESS',
  UPDATE_SUPPLIER_FAILURE = '[SUPPLIERS] UPDATE_SUPPLIER_FAILURE',
  DELETE_SUPPLIER = '[SUPPLIERS] DELETE_SUPPLIER',
  DELETE_SUPPLIER_SUCCESS = '[SUPPLIERS] DELETE_SUPPLIER_SUCCESS',
  DELETE_SUPPLIER_FAILURE = '[SUPPLIERS] DELETE_SUPPLIER_FAILURE',
}